Factor out orientation API operation
